Extract common fusion logic into config_utils
