+++ bcbio/utils.py
@@ -18,6 +18,7 @@ import types
 
 import toolz as tz
 import yaml
+
 try:
     from concurrent import futures
 except ImportError:
